requirements.txt
ansible
markupsafe
jinja2
PyYAML
